Complete rewrite of StickyHeaderLayoutManager
Fixes a rare problem I was seeing where wiggling the view rapidly could
cause views to not be rendered. Rewrite takes a different approach to
state and is simpler code. Instead of maintaining a “mirror state” of
the header/item/footer state, we now just  generate that info on demand
using view tag info.